pyproject.toml
[tool.poetry]
name = "ctk-color-picker-alpha"
version = "0.8.6"
description = "An eight-digits hex code color picker, with alpha value."
authors = ["Kolyn Lin <Kanzakimakai27875628@gmail.com>"]
license = "CC0 1.0 Universal"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
customtkinter = "5.2.2"
numpy = "1.26.4"
pillow = "10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
